// SistemaVentasBatia/ClientApp/src/app/models/productosumco.ts
export interface productossumco {
    idProducto: number;
    clave: string;
    nombre: string;
    idUnidad: number;
    unidad: string;
    precio: number;
    precioCompra: number;
    fechaAlta: string;
    personalAlta: number;
    fechaMod: string;
    personalMod: number;
    idEstatus: number;
}